+++ Source/DatabaseProviders/DatabaseProviderBase.cs
@@ -29,4 +29,6 @@ internal abstract class DatabaseProviderBase : IDatabaseProvider
 	public abstract void              ClearAllPools      (string providerName        );
 	public abstract DateTime?         GetLastSchemaUpdate(ConnectionSettings settings);
 	public abstract DbProviderFactory GetProviderFactory (string providerName        );
+
+	public virtual IEnumerable<(Type type, TypeRenderer renderer)> GetTypeRenderers() => Array.Empty<(Type, TypeRenderer)>();
 }