+++ Source/DatabaseProviders/DatabaseProviders.cs
@@ -6,36 +6,47 @@ namespace LinqToDB.LINQPad;
 
 internal static class DatabaseProviders
 {
+	private static readonly IReadOnlyDictionary<Type, TypeRenderer> _typeRenderers;
+
 	public static readonly IReadOnlyDictionary<string, IDatabaseProvider> Providers;
 	public static readonly IReadOnlyDictionary<string, IDatabaseProvider> ProvidersByProviderName;
 
 	static DatabaseProviders()
 	{
+		var typeRenderers       = new Dictionary<Type, TypeRenderer>();
 		var providers           = new Dictionary<string, IDatabaseProvider>();
 		var providersByName     = new Dictionary<string, IDatabaseProvider>();
 		Providers               = providers;
 		ProvidersByProviderName = providersByName;
+		_typeRenderers          = typeRenderers;
 
-		Register(providers, providersByName, new AccessProvider    ());
-		Register(providers, providersByName, new FirebirdProvider  ());
-		Register(providers, providersByName, new MySqlProvider     ());
-		Register(providers, providersByName, new PostgreSQLProvider());
-		Register(providers, providersByName, new SybaseAseProvider ());
-		Register(providers, providersByName, new SQLiteProvider    ());
-		Register(providers, providersByName, new SqlCeProvider     ());
-		Register(providers, providersByName, new DB2Provider       ());
-		Register(providers, providersByName, new InformixProvider  ());
-		Register(providers, providersByName, new SapHanaProvider   ());
-		Register(providers, providersByName, new OracleProvider    ());
-		Register(providers, providersByName, new SqlServerProvider ());
-		Register(providers, providersByName, new ClickHouseProvider());
+		Register(providers, providersByName, typeRenderers, new AccessProvider    ());
+		Register(providers, providersByName, typeRenderers, new FirebirdProvider  ());
+		Register(providers, providersByName, typeRenderers, new MySqlProvider     ());
+		Register(providers, providersByName, typeRenderers, new PostgreSQLProvider());
+		Register(providers, providersByName, typeRenderers, new SybaseAseProvider ());
+		Register(providers, providersByName, typeRenderers, new SQLiteProvider    ());
+		Register(providers, providersByName, typeRenderers, new SqlCeProvider     ());
+		Register(providers, providersByName, typeRenderers, new DB2Provider       ());
+		Register(providers, providersByName, typeRenderers, new InformixProvider  ());
+		Register(providers, providersByName, typeRenderers, new SapHanaProvider   ());
+		Register(providers, providersByName, typeRenderers, new OracleProvider    ());
+		Register(providers, providersByName, typeRenderers, new SqlServerProvider ());
+		Register(providers, providersByName, typeRenderers, new ClickHouseProvider());
 
-		static void Register(Dictionary<string, IDatabaseProvider> providers, Dictionary<string, IDatabaseProvider> providersByName, IDatabaseProvider provider)
+		static void Register(
+			Dictionary<string, IDatabaseProvider> providers,
+			Dictionary<string, IDatabaseProvider> providersByName,
+			Dictionary<Type  , TypeRenderer     > typeRenderers,
+			IDatabaseProvider provider)
 		{
 			providers.Add(provider.Database, provider);
 
 			foreach (var info in provider.Providers)
 				providersByName.Add(info.Name, provider);
+
+			foreach (var (type, renderer) in provider.GetTypeRenderers())
+				typeRenderers.Add(type, renderer);
 		}
 	}
 
@@ -78,5 +89,13 @@ internal static class DatabaseProviders
 			return provider;
 
 		throw new LinqToDBLinqPadException($"Cannot find provider for database '{database}'");
+	}
+
+	public static void RenderValue(ref object? value)
+	{
+		if (value != null && _typeRenderers.TryGetValue(value.GetType(), out var renderer))
+			renderer(ref value);
+
+		value = ValueFormatter.Format(value);
 	}
 }