+++ Source/DatabaseProviders/TypeRenderer.cs
@@ -1,0 +1,6 @@
+﻿namespace LinqToDB.LINQPad;
+
+/// <summary>
+/// Delegate for provider-specific type value rendering.
+/// </summary>
+internal delegate void TypeRenderer(ref object? value);